Improving browser targets test coverage
To check that we definitely change the transpilation based on the app's config/targets.js (we do, the bug I'm hunting didn't turn out to be here).
